api/requirements.txt
flask
moviepy
numpy
opencv-python
Werkzeug
torch
open-clip-torch
google-cloud-vision
ultralytics
Pillow